\name{gclust.kde}
\alias{gclust.kde}
\concept{Graph Clustering}
\title{Graph Clustering Using NMF}
\description{Performs Clustering of Graphs using KDE via embedding.  --
orginally called computeS }
\usage{
gclust.kde(glist,r,maxsvt=10) 
}
\arguments{
    \item{X.ist}{}
    \item{sigma}{}
}
\value{
}

\author{Minh Tang \email{mtang10@jhu.edu}}

\examples{
}
