\name{kernel.stat}
\alias{kernel.stat}
\concept{Graph Clustering}
\title{Graph Clustering Using KDE -- helper function }
\description{Help to Performs Clustering of Graphs using KDE via embedding.}
\usage{
kernel.stat(X,Y,sigma)
}
\arguments{
    \item{X}{}
    \item{Y}{}
    \item{sigma}{}
}
\value{
}

\author{Minh Tang \email{mtang10@jhu.edu}}

\examples{
}
