\name{plotmemb}
\alias{plotmemb}
\concept{Graph Visualization}
\title{Graph Visualization}
\description{Performs Visualization of Graphs 
\usage{
plotmemb(glist,r) 
}
\arguments{
    \item{X}
    \item{cl}
    \item{inorder}
    \item{main}
    \item{drawborder}
    \item{lwd}
    \item{lwdb}
    \item{lcol}
}
\value{
    No retutrn value, but a plot will be drawn in a new window
}

\author{Youngser Park \email{parky@jhu.edu}}

\examples{
}
