\name{rect.dist}
\alias{rect.dist}
\concept{Graph Clustering}
\title{Graph Clustering Using KDE -- helper function }
\description{Help to Performs Clustering of Graphs using KDE via embedding.}
\usage{
rect.dist(X,Y)
}
\arguments{
    \item{X}{}
    \item{Y}{}
}
\value{
}

\author{Minh Tang \email{mtang10@jhu.edu}}

\examples{
}
