\name{reembed}
\alias{reembed}
\concept{Graph Clustering}
\title{Graph Clustering Using KDE -- helper function }
\description{Help to Performs Clustering of Graphs using KDE via embedding.}
\usage{
reembed(g,dmax,pamkout) 
}
\arguments{
    \item{g}{}
    \item{dmax}{}
    \item{pamkout}{}
}
\value{
}

\author{Minh Tang \email{mtang10@jhu.edu}}

\examples{
}
